Broke tests out
* Broke unit/smoke tests into directories
* Cleaned up specs to be more mocha-ish and match the proper
  js testing style
* updated watcher script
* increased timeouts on tests because jankins is slow ;_;
* moved zk test to integration dir, still need a good story here
* various jshint fixes where I see them
