+++ lib/config/test.js
@@ -6,7 +6,7 @@ export default {
   },
 
   log : {
-    level: 'error',
+    level: 'critical',
     name : 'geo-import'
   }
