+++ lib/decoders/disk.js
@@ -13,8 +13,7 @@ class Disk {
   _cleanup() {
     this._allocations.forEach((name) => {
       var warn = () => {
-        //TODO log this
-        return `Failed to clean up ${name}`;
+        return logger.warn(`Failed to clean up ${name}`)
       };
 
       fs.stat(name, (err, stat) => {
