+++ lib/decoders/kmz.js
@@ -40,7 +40,7 @@ class KMZ extends Duplex {
 
   _write(chunk, encoding, done) {
     logger.debug('Writing kmz chunk')
-    return this._zBuffer.write(new Buffer(chunk.toString('binary'), 'binary'), 'binary', done);
+    return this._zBuffer.write(chunk, null, done);
   }
 
   _onFinished() {
