+++ lib/decoders/shapefile.js
@@ -42,6 +42,7 @@ class Shapefile extends Duplex {
     this._fgroup = uuid.v4();
     this._zName = this._fileGroup('.zip');
     this._zBuffer = this._disk.allocate(this._zName, {
+      encoding: null,
       defaultEncoding: 'binary'
     });
 
@@ -59,8 +60,8 @@ class Shapefile extends Duplex {
   }
 
   _write(chunk, encoding, done) {
-    logger.debug(`Writing shape feature chunk ${typeof chunk} with ${encoding} encoding`);
-    return this._zBuffer.write(new Buffer(chunk.toString('binary'), 'binary'), 'binary', done);
+    logger.debug(`Writing shape feature chunk`);
+    return this._zBuffer.write(chunk, null, done);
   }
 
 
@@ -146,6 +147,9 @@ class Shapefile extends Duplex {
             if (err) return this.emit('error', err);
             this.emit('end');
           });
+        })
+        .on('error', (err) => {
+          this.emit('error', err);
         });
 
     }.bind(this));
