+++ lib/services/spatial.js
@@ -67,9 +67,9 @@ class SpatialService {
     return async.map(upsertSpecs, core.upsert.bind(core), (err, upsert) => {
       if(err) return res.status(503).send(err);
 
-      return async.map(upsert, ([[layer, fourfour], upsertRequest], cb) => {
+      return async.map(upsert, ([[layer, fourfour], upsertBuilder], cb) => {
         req.log.info(`Starting upsert to ${fourfour}`);
-        layer.pipe(upsertRequest)
+        layer.pipe(upsertBuilder())
           .on('response', (upsertResponse) => {
             req.log.info(`Got upsert response ${upsertResponse}`);
             //bb hack to match the rest of the flow
@@ -96,7 +96,7 @@ class SpatialService {
 
       }, (err, upsertResponses) => {
 
-        err && req.log.info(`Upsert response error: ${err}`);
+        err && req.log.error(`Upsert response error: ${err} ${err.stack}`);
         if (err) return res.status(err.statusCode || 503).send(err.body);
 
         var layerResults = upsertResponses.map(([
