+++ lib/upstream/core.js
@@ -49,7 +49,7 @@ class Core {
   //partial to buffer a response
   bufferResponse(onBuffered) {
     return (response) => {
-      if(!response.pipe) {
+      if (!response.pipe) {
         //this is so gross.
         //the error even will emit both error responses
         //(shouldn't 'response' do that?) as well as regular
@@ -65,22 +65,22 @@ class Core {
       }
 
       response.pipe(reduceStream((acc, data) => {
-        return acc + data.toString('utf-8');
-      }, ''))
-      .on('data', (buf) => {
-        try {
-          response.body = JSON.parse(buf);
-        } catch(e) {
-          response.body = buf;
-        }
-        onBuffered(response);
-      });
+          return acc + data.toString('utf-8');
+        }, ''))
+        .on('data', (buf) => {
+          try {
+            response.body = JSON.parse(buf);
+          } catch (e) {
+            response.body = buf;
+          }
+          onBuffered(response);
+        });
     };
   }
 
   _onResponseStart(onComplete) {
     return this.bufferResponse((response) => {
-      if(response.statusCode > 300) return onComplete(response);
+      if (response.statusCode > 300) return onComplete(response);
       return onComplete(false, response);
     });
   }
@@ -96,22 +96,22 @@ class Core {
       if (err) return onComplete(err);
 
       request.post({
-        url: `${url}/views?nbe=true`,
-        headers: this._headers(),
-        body: {
-          name: layer.name
-        },
-        json: true
-      })
-      .on('response', this._onResponseStart(onComplete))
-      .on('error', this._onErrorResponse(onComplete));
+          url: `${url}/views?nbe=true`,
+          headers: this._headers(),
+          body: {
+            name: layer.name
+          },
+          json: true
+        })
+        .on('response', this._onResponseStart(onComplete))
+        .on('error', this._onErrorResponse(onComplete));
     }.bind(this));
   }
 
   addColumn(colSpec, onComplete) {
     logger.info(`Add column ${colSpec} to core`);
     var [fourfour, [name, dataTypeName]] = colSpec;
-    if(!name || !dataTypeName) {
+    if (!name || !dataTypeName) {
       return onComplete("Invalid column, name and dataTypeName cannot be null!");
     }
 
@@ -119,17 +119,17 @@ class Core {
       if (err) return onComplete(err);
 
       return request.post({
-        url: `${url}/views/${fourfour}/columns`,
-        headers: this._headers(),
-        body: {
-          fieldName: name,
-          name: name,
-          dataTypeName: dataTypeName
-        },
-        json: true
-      })
-      .on('response', this._onResponseStart(onComplete))
-      .on('error', this._onErrorResponse(onComplete));
+          url: `${url}/views/${fourfour}/columns`,
+          headers: this._headers(),
+          body: {
+            fieldName: name,
+            name: name,
+            dataTypeName: dataTypeName
+          },
+          json: true
+        })
+        .on('response', this._onResponseStart(onComplete))
+        .on('error', this._onErrorResponse(onComplete));
 
     }.bind(this));
   }
@@ -141,10 +141,13 @@ class Core {
     logger.info(`Upsert to core ${fourfour}`);
     return this._url((err, url) => {
       if (err) return onOpened(err);
-      return onOpened(false, [upsertSpec, request.post({
-        url: `${url}/id/${fourfour}.json`,
-        headers: this._headers()
-      })]);
+      var upsertOpener = () => {
+        return request.post({
+          url: `${url}/id/${fourfour}.json`,
+          headers: this._headers()
+        });
+      };
+      return onOpened(false, [upsertSpec, upsertOpener]);
     });
   }
 }
