--- test/test.js
@@ -1,5 +1,0 @@
-'use strict';
-
-var chai = require('chai');
-var expect = chai.expect;
-
